client: move callback typedefs and arg struct into ceph_ll_client.h
Put them in a common interface header file. This also allows us to
eliminate the duplicate definition of ceph_deleg_cb_t in Delegation.h.

(cherry picked from commit 8370f70cacfb0bccc96d121cab687376155e8b8d)
